Update README.md
Semantic HTML is crucial for programmers because it improves user experience on websites by offering clear navigational structures like screen readers and signposts. It also helps search engines accurately interpret and classify your content, which improves rankings and increases online visibility. Furthermore, because it adheres to a consistent and logical structure, semantic HTML is easier to create, maintain, and comprehend